Fix button 4 not activating
UE is a perfect game engine without any flaws or anything like that.
Not in the least.

That is why button 4 does not activate if you spawn pawns on top of it.
They get stuck in the platform above and don't collide with anything.
However, if you spawn a pawn somewhere completely different, and move it
on top of button 4 literally in the same tick in the same frame as the
very next instruction to the game engine, it still gets stuck, but for
some reason activates the button.